+++ src/intellilightssimulator/Hardware/EnergyPackage/SolarPanel.java
@@ -11,19 +11,19 @@ package intellilightssimulator.Hardware.EnergyPackage;
  */
 public class SolarPanel {
     
-    private final float Vmpp;
-    private final float Impp;
+    private final double Vmpp;
+    private final double Impp;
     private final int XDimensions;
     private final int YDimensions;
     
-    public SolarPanel(float Vmpp, float Impp, int XDimensions, int YDimensions ){
+    public SolarPanel(double Vmpp, double Impp, int XDimensions, int YDimensions ){
         this.Vmpp = Vmpp;
         this.Impp = Impp;
         this.XDimensions = XDimensions;
         this.YDimensions = YDimensions;
     }
     
-    public int getCharge(int startTime, int endTime) {
+    public double getCharge(int startTime, int endTime) {
         return 1; //calculation for time-period here.
     }
 }
